UKFetcher -> UKEAFetcher, add metadata support
